# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VID Master Zimbabwe API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tests/start:
    post:
      operationId: startTest
      tags: [tests]
      summary: Start a new timed test session (25 questions, 8 min)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartTestInput"
      responses:
        "201":
          description: Test session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestSession"
  /tests/{sessionId}/submit:
    post:
      operationId: submitTest
      tags: [tests]
      summary: Submit test answers and get results
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestSubmission"
      responses:
        "200":
          description: Test results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestResult"
  /tests/history:
    get:
      operationId: getTestHistory
      tags: [tests]
      summary: Get current user's test history
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: offset
          schema:
            type: integer
      responses:
        "200":
          description: Test history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TestResult"
  /tests/{sessionId}:
    get:
      operationId: getTestResult
      tags: [tests]
      summary: Get a specific test result by session ID
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestResult"

  # --- PROGRESS ---
...
components:
  schemas:
...
    Question:
      type: object
      required: [id, text, options, correctAnswer, category, difficulty, explanation]
      properties:
        id:
          type: integer
        text:
          type: string
        options:
          type: array
          items:
            type: string
        correctAnswer:
          type: integer
        category:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        explanation:
          type: string
        imageUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, approved, published, archived]
        createdAt:
          type: string
...
    StartTestInput:
      type: object
      properties:
        mode:
          type: string
          enum: [timed, practice]
    AnswerInput:
      type: object
      required: [questionId, selectedAnswer]
      properties:
        questionId:
          type: integer
        selectedAnswer:
          type: integer
    TestSubmission:
      type: object
      required: [answers]
      properties:
        answers:
          type: array
          items:
            $ref: "#/components/schemas/AnswerInput"
    TestSession:
      type: object
      required: [sessionId, questions, startedAt, expiresAt, durationSeconds]
      properties:
        sessionId:
          type: string
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        startedAt:
          type: string
        expiresAt:
          type: string
        durationSeconds:
          type: integer
        mode:
          type: string
    AnswerDetail:
      type: object
      required: [questionId, selectedAnswer, correctAnswer, isCorrect]
      properties:
        questionId:
          type: integer
        text:
          type: string
        selectedAnswer:
          type: integer
        correctAnswer:
          type: integer
        isCorrect:
          type: boolean
        explanation:
          type: string
    TestResult:
      type: object
      required: [sessionId, score, total, percentage, passed, timeTaken, completedAt]
      properties:
        sessionId:
          type: string
        score:
          type: integer
        total:
          type: integer
        percentage:
          type: number
        passed:
          type: boolean
        timeTaken:
          type: integer
        completedAt:
          type: string
        mode:
          type: string
        answers:
          type: array
          items:
            $ref: "#/components/schemas/AnswerDetail"

    # Progress
...